move format conversion out of sensor and make D405 get_profile_tags() aware of it
